+++ astra-sim/system/collective/DoubleBinaryTreeAllReduce.hh
@@ -50,7 +50,7 @@ class DoubleBinaryTreeAllReduce : public Algorithm {
       int id,
       int layer_num,
       BinaryTree* tree,
-      int data_size,
+      uint64_t data_size,
       bool boost_mode);
   // void init(BaseStream *stream);
 };
