+++ astra-sim/system/collective/HalvingDoubling.hh
@@ -48,7 +48,7 @@ namespace AstraSim {
         long free_packets;
         long total_packets_sent;
         long total_packets_received;
-        int msg_size;
+        uint64_t msg_size;
         std::list<MyPacket*> locked_packets;
         bool processed;
         bool send_back;
@@ -63,7 +63,7 @@ namespace AstraSim {
                 int id,
                 int layer_num,
                 RingTopology* ring_topology,
-                int data_size,
+                uint64_t data_size,
                 bool boost_mode);
         virtual void run(EventType event, CallData* data);
         RingTopology::Direction specify_direction();
